Se implementa la funcionalidad de vista alterna en la página de favoritos, permitiendo alternar entre vista de cuadrícula y lista. Se simplifica la página de detalles al utilizar un componente DetailCard para mostrar la información del Pokémon.
